requirements.txt
Flask==3.0.0
TensorFlow==2.15.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
Werkzeug==3.0.1

